+++ include/viua/bytecode/maps.h
@@ -119,7 +119,7 @@ const std::map<std::string, unsigned> OP_SIZES = {
 
     { "throw",          sizeof(byte) + sizeof(OperandType) + sizeof(int) },
     { "catch",          sizeof(byte) }, // catch "<type>" <block>
-    { "pull",           sizeof(byte) + sizeof(OperandType) + sizeof(int) }, // pull <register>
+    { "draw",           sizeof(byte) + sizeof(OperandType) + sizeof(int) }, // draw <register>
     { "try",            sizeof(byte) },
     { "enter",          sizeof(byte) },
     { "leave",          sizeof(byte) },
@@ -231,7 +231,7 @@ const std::map<enum OPCODE, std::string> OP_NAMES = {
 
     { THROW,        "throw" },
     { CATCH,        "catch" },
-    { PULL,         "pull" },
+    { DRAW,         "draw" },
     { TRY,          "try" },
     { ENTER,        "enter" },
     { LEAVE,        "leave" },